Correcciones de Greedy
