+++ src/tpe/Greedy.java
@@ -1,6 +1,7 @@
 package ProgramacionIII.tpe;
 
 import java.util.ArrayList;
+import java.util.Comparator;
 import java.util.List;
 
 public class Greedy {
@@ -15,7 +16,26 @@ public class Greedy {
         this.procesadores = new ArrayList<>(procesadores);
         this.estadosGenerados = 0;
         this.resulTiempoFinalEjecucion = 0;
+        this.tareas.sort(comparadorTiempoEjecucion);
     }
+
+    /*
+    Nosotros consideramos que era una buena estrategia ordenar por criticas y tiempo de ejecucion(mayor a menor)
+    para encontrar los casos de procesadores saturados mas rapido y asi descartar estados innecesario.
+     */
+    Comparator<Tarea> comparadorTiempoEjecucion = new Comparator<Tarea>() {
+        @Override
+        public int compare(Tarea tarea1, Tarea tarea2) {
+            // Si las tareas son críticas, pone la crítica primero
+            if (tarea1.getEsCritica() && !tarea2.getEsCritica()) {
+                return -1;
+            } else if (!tarea1.getEsCritica() && tarea2.getEsCritica()) {
+                return 1;
+            }
+            // Si ambas tareas son críticas o no críticas, ordena por tiempo de ejecucion de mayor a menor
+            return Integer.compare(tarea2.getTiempoEjecucion(), tarea1.getTiempoEjecucion());
+        }
+    };
 
     /*
     Este metodo es nuestro criterio para seleccionar la mejor opcion posible, en este caso el procesador
