+++ src/tpe/Servicios.java
@@ -92,8 +92,6 @@ public class Servicios {
 
     public Solucion asginarTareasConBacktracking(Integer maxTiempoNoRefrigerados) {
         this.procesadores.stream().filter(p -> !p.getRefrigerado()).collect(Collectors.toList()).forEach(p -> p.setTiempoMaximo(maxTiempoNoRefrigerados));
-        tareasAsignar.sort(comparadorTiempoEjecucion);
-        System.out.println(tareasAsignar);
         Backtracking backtracking = new Backtracking(tareasAsignar, this.procesadores);
         return backtracking.iniciarBacktracking();
     }
@@ -110,29 +108,12 @@ public class Servicios {
 
     public Solucion asignarTareasConGreedy(Integer maxTiempoNoRefrigerados) {
         this.procesadores.stream().filter(p -> !p.getRefrigerado()).collect(Collectors.toList()).forEach(p -> p.setTiempoMaximo(maxTiempoNoRefrigerados));
-        tareasAsignar.sort(comparadorTiempoEjecucion);
         Greedy greedy = new Greedy(tareasAsignar, procesadores);
         return greedy.ejecutarGreedy();
 
     }
 
-    /*
-    Nosotros consideramos que era una buena estrategia ordenar por criticas y tiempo de ejecucion(mayor a menor)
-    para encontrar los casos de procesadores saturados mas rapido y asi descartar estados innecesario.
-     */
-    Comparator<Tarea> comparadorTiempoEjecucion = new Comparator<Tarea>() {
-        @Override
-        public int compare(Tarea tarea1, Tarea tarea2) {
-            // Si las tareas son críticas, pone la crítica primero
-            if (tarea1.getEsCritica() && !tarea2.getEsCritica()) {
-                return -1;
-            } else if (!tarea1.getEsCritica() && tarea2.getEsCritica()) {
-                return 1;
-            }
-            // Si ambas tareas son críticas o no críticas, ordena por tiempo de ejecucion de mayor a menor
-            return Integer.compare(tarea2.getTiempoEjecucion(), tarea1.getTiempoEjecucion());
-        }
-    };
+
 
 
 }
